refactor: move exit code to its own module
Mostly to keep the number of lines in bookkeeper.py below the 1000 limit.
